+++ examples/linkproppred/redditcomments/edgebank_np.py
@@ -65,9 +65,10 @@ def test_one_vs_many(data, test_mask, neg_sampler, split_mode):
 
     return perf_metrics
 
+
 def get_args():
     parser = argparse.ArgumentParser('*** TGB: EdgeBank ***')
-    parser.add_argument('-d', '--data', type=str, help='Dataset name', default='wikipedia')
+    parser.add_argument('-d', '--data', type=str, help='Dataset name', default='redditcomments')
     parser.add_argument('--bs', type=int, help='Batch size', default=200)
     parser.add_argument('--k_value', type=int, help='k_value for computing ranking metrics', default=10)
     parser.add_argument('--seed', type=int, help='Random seed', default=1)
