+++ examples/linkproppred/stablecoin/edgebank_np.py
@@ -67,7 +67,7 @@ def test_one_vs_many(data, test_mask, neg_sampler, split_mode):
 
 def get_args():
     parser = argparse.ArgumentParser('*** TGB: EdgeBank ***')
-    parser.add_argument('-d', '--data', type=str, help='Dataset name', default='wikipedia')
+    parser.add_argument('-d', '--data', type=str, help='Dataset name', default='stablecoin')
     parser.add_argument('--bs', type=int, help='Batch size', default=200)
     parser.add_argument('--k_value', type=int, help='k_value for computing ranking metrics', default=10)
     parser.add_argument('--seed', type=int, help='Random seed', default=1)
