+++ examples/linkproppred/stablecoin/tgn_early_stop.py
