--- modules/tgn_memory.py
@@ -1,215 +1,0 @@
-"""
-Memory Module
-
-Reference:
-    - https://pytorch-geometric.readthedocs.io/en/latest/_modules/torch_geometric/nn/models/tgn.html
-"""
-
-
-import copy
-from typing import Callable, Dict, Tuple
-
-import torch
-from torch import Tensor
-from torch.nn import GRUCell, RNNCell, Linear
-
-from torch_geometric.nn.inits import zeros
-from torch_geometric.utils import scatter
-
-from modules.time_enc import TimeEncoder
-
-
-TGNMessageStoreType = Dict[int, Tuple[Tensor, Tensor, Tensor, Tensor]]
-
-
-class TGNMemory(torch.nn.Module):
-    r"""The Temporal Graph Network (TGN) memory model from the
-    `"Temporal Graph Networks for Deep Learning on Dynamic Graphs"
-    <https://arxiv.org/abs/2006.10637>`_ paper.
-
-    .. note::
-
-        For an example of using TGN, see `examples/tgn.py
-        <https://github.com/pyg-team/pytorch_geometric/blob/master/examples/
-        tgn.py>`_.
-
-    Args:
-        num_nodes (int): The number of nodes to save memories for.
-        raw_msg_dim (int): The raw message dimensionality.
-        memory_dim (int): The hidden memory dimensionality.
-        time_dim (int): The time encoding dimensionality.
-        message_module (torch.nn.Module): The message function which
-            combines source and destination node memory embeddings, the raw
-            message and the time encoding.
-        aggregator_module (torch.nn.Module): The message aggregator function
-            which aggregates messages to the same destination into a single
-            representation.
-    """
-
-    def __init__(
-        self,
-        num_nodes: int,
-        raw_msg_dim: int,
-        memory_dim: int,
-        time_dim: int,
-        message_module: Callable,
-        aggregator_module: Callable,
-        memory_updater_cell: str = "gru",
-    ):
-        super().__init__()
-
-        self.num_nodes = num_nodes
-        self.raw_msg_dim = raw_msg_dim
-        self.memory_dim = memory_dim
-        self.time_dim = time_dim
-
-        self.msg_s_module = message_module
-        self.msg_d_module = copy.deepcopy(message_module)
-        self.aggr_module = aggregator_module
-        self.time_enc = TimeEncoder(time_dim)
-        # self.gru = GRUCell(message_module.out_channels, memory_dim)
-        if memory_updater_cell == "gru":  # for TGN
-            self.memory_updater = GRUCell(message_module.out_channels, memory_dim)
-        elif memory_updater_cell == "rnn":  # for JODIE & DyRep
-            self.memory_updater = RNNCell(message_module.out_channels, memory_dim)
-        else:
-            raise ValueError(
-                "Undefined memory updater!!! Memory updater can be either 'gru' or 'rnn'."
-            )
-
-        self.register_buffer("memory", torch.empty(num_nodes, memory_dim))
-        last_update = torch.empty(self.num_nodes, dtype=torch.long)
-        self.register_buffer("last_update", last_update)
-        self.register_buffer("_assoc", torch.empty(num_nodes, dtype=torch.long))
-
-        self.msg_s_store = {}
-        self.msg_d_store = {}
-
-        self.reset_parameters()
-
-    @property
-    def device(self) -> torch.device:
-        return self.time_enc.lin.weight.device
-
-    def reset_parameters(self):
-        r"""Resets all learnable parameters of the module."""
-        if hasattr(self.msg_s_module, "reset_parameters"):
-            self.msg_s_module.reset_parameters()
-        if hasattr(self.msg_d_module, "reset_parameters"):
-            self.msg_d_module.reset_parameters()
-        if hasattr(self.aggr_module, "reset_parameters"):
-            self.aggr_module.reset_parameters()
-        self.time_enc.reset_parameters()
-        self.memory_updater.reset_parameters()
-        self.reset_state()
-
-    def reset_state(self):
-        """Resets the memory to its initial state."""
-        zeros(self.memory)
-        zeros(self.last_update)
-        self._reset_message_store()
-
-    def detach(self):
-        """Detaches the memory from gradient computation."""
-        self.memory.detach_()
-
-    def forward(self, n_id: Tensor) -> Tuple[Tensor, Tensor]:
-        """Returns, for all nodes :obj:`n_id`, their current memory and their
-        last updated timestamp."""
-        if self.training:
-            memory, last_update = self._get_updated_memory(n_id)
-        else:
-            memory, last_update = self.memory[n_id], self.last_update[n_id]
-
-        return memory, last_update
-
-    def update_state(self, src: Tensor, dst: Tensor, t: Tensor, raw_msg: Tensor):
-        """Updates the memory with newly encountered interactions
-        :obj:`(src, dst, t, raw_msg)`."""
-        n_id = torch.cat([src, dst]).unique()
-
-        if self.training:
-            self._update_memory(n_id)
-            self._update_msg_store(src, dst, t, raw_msg, self.msg_s_store)
-            self._update_msg_store(dst, src, t, raw_msg, self.msg_d_store)
-        else:
-            self._update_msg_store(src, dst, t, raw_msg, self.msg_s_store)
-            self._update_msg_store(dst, src, t, raw_msg, self.msg_d_store)
-            self._update_memory(n_id)
-
-    def _reset_message_store(self):
-        i = self.memory.new_empty((0,), device=self.device, dtype=torch.long)
-        msg = self.memory.new_empty((0, self.raw_msg_dim), device=self.device)
-        # Message store format: (src, dst, t, msg)
-        self.msg_s_store = {j: (i, i, i, msg) for j in range(self.num_nodes)}
-        self.msg_d_store = {j: (i, i, i, msg) for j in range(self.num_nodes)}
-
-    def _update_memory(self, n_id: Tensor):
-        memory, last_update = self._get_updated_memory(n_id)
-        self.memory[n_id] = memory
-        self.last_update[n_id] = last_update
-
-    def _get_updated_memory(self, n_id: Tensor) -> Tuple[Tensor, Tensor]:
-        self._assoc[n_id] = torch.arange(n_id.size(0), device=n_id.device)
-
-        # Compute messages (src -> dst).
-        msg_s, t_s, src_s, dst_s = self._compute_msg(
-            n_id, self.msg_s_store, self.msg_s_module
-        )
-
-        # Compute messages (dst -> src).
-        msg_d, t_d, src_d, dst_d = self._compute_msg(
-            n_id, self.msg_d_store, self.msg_d_module
-        )
-
-        # Aggregate messages.
-        idx = torch.cat([src_s, src_d], dim=0)
-        msg = torch.cat([msg_s, msg_d], dim=0)
-        t = torch.cat([t_s, t_d], dim=0)
-        aggr = self.aggr_module(msg, self._assoc[idx], t, n_id.size(0))
-
-        # Get local copy of updated memory.
-        memory = self.memory_updater(aggr, self.memory[n_id])
-
-        # Get local copy of updated `last_update`.
-        dim_size = self.last_update.size(0)
-        last_update = scatter(t, idx, 0, dim_size, reduce="max")[n_id]
-
-        return memory, last_update
-
-    def _update_msg_store(
-        self,
-        src: Tensor,
-        dst: Tensor,
-        t: Tensor,
-        raw_msg: Tensor,
-        msg_store: TGNMessageStoreType,
-    ):
-        n_id, perm = src.sort()
-        n_id, count = n_id.unique_consecutive(return_counts=True)
-        for i, idx in zip(n_id.tolist(), perm.split(count.tolist())):
-            msg_store[i] = (src[idx], dst[idx], t[idx], raw_msg[idx])
-
-    def _compute_msg(
-        self, n_id: Tensor, msg_store: TGNMessageStoreType, msg_module: Callable
-    ):
-        data = [msg_store[i] for i in n_id.tolist()]
-        src, dst, t, raw_msg = list(zip(*data))
-        src = torch.cat(src, dim=0)
-        dst = torch.cat(dst, dim=0)
-        t = torch.cat(t, dim=0)
-        raw_msg = torch.cat(raw_msg, dim=0)
-        t_rel = t - self.last_update[src]
-        t_enc = self.time_enc(t_rel.to(raw_msg.dtype))
-
-        msg = msg_module(self.memory[src], self.memory[dst], raw_msg, t_enc)
-
-        return msg, t, src, dst
-
-    def train(self, mode: bool = True):
-        """Sets the module in training mode."""
-        if self.training and not mode:
-            # Flush message store to memory in case we just entered eval mode.
-            self._update_memory(torch.arange(self.num_nodes, device=self.memory.device))
-            self._reset_message_store()
-        super().train(mode)
